server: eliminate handshake process from session management
This intend to eliminate handshake process from session management,
therefore we do not need to do a handshake when we join a new user.
In this case, the user goroutine automatically set the parameters that requires
handshake to processed.
At the moment, we use C helper function to set default parameters rather than
entirely reimplement libguac.
Since we simplified handshake in libguac, connection.go and session.go are merged together.

See #2.